fix team-record
I was struggling for a long while to make this fast.

First I tried adding extra clauses onto the query and was confused at
how much slower they made everything.

Then I made get-name-map to push this data into clojure's memory

Both of these were very slow.

The main issue ended up being that a query with playerID first will be
slow; it needs to load in all of the player information.  This can be
fixed in team-record simply by putting the playerID query below the
teamID and playerID filters.

get-name-map is still around in case we need it.  It's understandably
very slow